+++ src/main/java/com/fontys/api/repositories/UserRepository.java
@@ -3,6 +3,6 @@ package com.fontys.api.repositories;
 import com.fontys.api.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long>
+public interface UserRepository extends JpaRepository<User, Integer>
 {
 }
